Convert CsMonad into a newtype.
The underlying reason here was to make type error messages easier to
understand because it would show the name of the type instead of its
component parts.

However, doing this requires that I then make CsMonad an instance of a
lot of type classes.  Making this work with the bdcs repo requires
further instances, but there's no time to do that now.  I'm not
confident these are correct.  Tests would go a long way towards
convincing me.